# Replace `internal::GRAIN_SIZE` by `grain_size` (parameter).