Fixes #53013